requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
csv-to-sqlite==2.1.3
Django==3.2
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
idna==3.4
iniconfig==2.0.0
numpy==1.24.2
packaging==23.0
pandas==2.0.0
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2023.3
urllib3==1.26.15
